+++ Source/Application/Receiver.h
@@ -149,8 +149,8 @@ public:
 	std::unique_ptr<AIS::Model> &addModel(int m);
 	std::unique_ptr<AIS::Model> &Model(int i) { return models[i]; }
 	int Count() { return models.size(); }
-	uint32_t getGroupMask() const {
-		uint32_t mask = 0;
+	uint64_t getGroupMask() const {
+		uint64_t mask = 0;
 		for (int i = 0; i < (int)jsonais.size(); i++)
 			mask |= jsonais[i]->out.getGroupOut();
 		return mask;
